Move Total sensors to sensor.py
